UserStory 7.1: Refactored large parts of the code. Made it so there are no crashes if the user inputs invalid variables such as a string where there should be an integer. Moved duplicate code into a separate method.
Added a bunch of sc.nextLine() in order to clear input streams after nextInt() methods.

Need to refactor the edit case as it is very messy and long. Best case is to break it into different methods.

Ask teacher if it's legal for UserInterface to create movie objects and call their getter/setters.