Added the ability to remove movie objects from the MovieCollection. Added try catches at critical points of failure.
TODO:
Change it so that edit, add or remove a movie is in a sub-menu after having been shown the list of movies.

Add it so that when you search for a movie and find it you can choose to edit or remove it instantly.

Still need to refactor UI so it looks less cluttered.